+++ remoulade/worker.py
@@ -405,7 +405,7 @@ class _WorkerThread(Thread):
             if isinstance(e, RateLimitExceeded):
                 self.logger.warning("Rate limit exceeded in message %s: %s.", message, e)
             else:
-                self.logger.warning("Failed to process message %s with unhandled exception.", message, exc_info=True)
+                self.logger.error("Failed to process message %s with unhandled exception.", message, exc_info=True)
 
             self.broker.emit_after("process_message", message, exception=e)